Move functions from servicedetailsbuilder.go to registerservice.go
