+++ github-connector/internal/registration/servicedetailsbuilder_test.go
@@ -1,0 +1,1 @@
+package registration
